Add support for local_addr
It would be nice if it were possible to obtain the local address that
the manager is listening on, for cases of ephemeral ports.  This adds
that support, which has some threading implications, by creating a
new local_addr property and a get_local_addr() method (called by the
local_addr property getter) which allows alternate timeouts.