Add new features to SocketCloser
Two new features were added to SocketCloser.  The first is an error
threshold tracking, which allows for extreme simplification of the
accept loop in tcp.py.  The second is a capability to ignore certain
exceptions; a new application.RejectConnection exception is added
and subsequently ignored in the connect() and listener() methods of
TCPTendrilManager through this new feature.